re-deploy on Ubuntu22

+++ chainsaw-python/das/algo_spectrum.py
@@ -42,19 +42,18 @@ strain_rate_real, strain_rate_imag = get_phase_diff(
 )
 
 phase_scaling_factor = 1 << 13
-
 phase = np.arctan2(strain_rate_imag, strain_rate_real)
 phase = (phase * phase_scaling_factor).astype(np.int32)
 
-your_upper = np.fromfile("/home/ltr/SpinalHDL/upper_result.bin", dtype=np.int16).astype(np.float32)
-your_lower = np.fromfile("/home/ltr/SpinalHDL/lower_result.bin", dtype=np.int16).astype(np.float32)
-your_float = np.fromfile("/home/ltr/SpinalHDL/result_float32.bin", dtype=np.float32)
+
+your_upper = np.fromfile("../../upper_result.bin", dtype=np.int16).astype(np.float32)
+your_lower = np.fromfile("../../lower_result.bin", dtype=np.int16).astype(np.float32)
+your_float = np.fromfile("../../result_float32.bin", dtype=np.float32)
 
 import shutil
-shutil.rmtree('test')
+if os.path.exists('test'):
+    shutil.rmtree('test')
 os.mkdir('test')
-
-import numpy as np
 
 valid_range = range(5 * pulse_valid_points, 6 * pulse_valid_points)
 plot_time_and_frequency(np.floor(np.log2(your_float)) + 16 + 1, 500e6, 'effective_bits.png', plot_range=valid_range)
@@ -62,3 +61,7 @@ plot_time_and_frequency(strain_rate_real.flatten(), 500e6, 'your_strain_rate_rea
 plot_time_and_frequency(strain_rate_real.flatten(), 500e6, 'your_strain_rate_real.png', yours=your_float, plot_range=valid_range)
 plot_time_and_frequency(phase.flatten(), 500e6, 'your_phase.png', yours=your_lower, plot_range=valid_range)
 plot_time_and_frequency(phase.flatten(), 500e6, 'your_phase_all.png', yours=your_lower)
+
+
+# plot_waterfall(phase)
+# plot_waterfall(your_lower.reshape(pulse_count, -1))
